+++ voting/frontend/components/admin/WorkflowManager.tsx
@@ -1,15 +1,17 @@
 "use client"
 
-import {useState} from "react"
+import {useEffect, useState} from "react"
 import {readContractByFunctionName, writeContractByFunctionName} from "@/utils";
 import {useWorkflowStatusContext} from "@/context/workflowStatus";
 import VoterManager from "@/components/admin/VoterManager";
 import Event from "@/components/Event";
+import {useToast} from "@chakra-ui/react";
 
 const WorkflowManager = () => {
-    const [error, setError] = useState('')
+    // const [error, setError] = useState('')
     const [success, setSuccess] = useState('')
     const { workflowStatus, setWorkflowStatus} = useWorkflowStatusContext()
+    const toast = useToast()
 
     const WorkflowStatus: string[] = [
         "RegisteringVoters",
@@ -20,17 +22,32 @@ const WorkflowManager = () => {
         "VotesTallied"
     ]
 
+    useEffect(() => {
+        getWorkflowStatus()
+    }, [success])
+
     const updateWorkflowStatus = (funcName: string) => {
-        setError('')
         setSuccess('')
 
         writeContractByFunctionName(funcName).then(
             hash => {
                 setSuccess(hash)
-                getWorkflowStatus()
+                toast({
+                    title: 'Workflow status updated.',
+                    description: `with status: ${funcName}.`,
+                    status: 'success',
+                    duration: 5000,
+                    isClosable: true,
+                })
             }
         ).catch(
-            err => setError(err.message)
+            err => toast({
+                title: "Unable to update the workflow status",
+                description: err.message,
+                status: 'error',
+                duration: 5000,
+                isClosable: true,
+            })
         )
     }
 
@@ -44,60 +61,57 @@ const WorkflowManager = () => {
 
     return (
         <>
-            <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-600 text-indigo-100 shadow-lg">
-                <div className="p-4">
+            <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-800 text-indigo-100 shadow-lg drop-shadow-lg border-indigo-600 border">
+                <div className="p-4 font-bold text-md">
                     WorkflowStatus : {WorkflowStatus[workflowStatus]}
                 </div>
             </div>
 
-            {WorkflowStatus[workflowStatus] === "RegisteringVoters" ? (<VoterManager />) : (<></>)}
+            {WorkflowStatus[workflowStatus] === "RegisteringVoters" && <VoterManager /> }
 
-            {WorkflowStatus[workflowStatus] !== "VotesTallied" ? (
+            {WorkflowStatus[workflowStatus] !== "VotesTallied" &&
 
-                <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-600 text-indigo-100 shadow-lg">
+                <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-800 text-indigo-100 shadow-lg drop-shadow-lg border-indigo-600 border">
                     <div className="p-4">
 
-                        {WorkflowStatus[workflowStatus] === "RegisteringVoters" ? (
+                        {WorkflowStatus[workflowStatus] === "RegisteringVoters" &&
                             <button
                                 onClick={() => updateWorkflowStatus("startProposalsRegistering")}
-                                className="content-center bg-indigo-950 hover:bg-indigo-900 text-white font-semibold py-2 px-4 rounded-full"
+                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
                             >Start Proposals Registering</button>
-                        ) : (<></>)}
+                        }
 
-                        {WorkflowStatus[workflowStatus] === "ProposalsRegistrationStarted" ? (
+                        {WorkflowStatus[workflowStatus] === "ProposalsRegistrationStarted" &&
                             <button
                                 onClick={() => updateWorkflowStatus("endProposalsRegistering")}
-                                className="content-center bg-indigo-950 hover:bg-indigo-900 text-white font-semibold py-2 px-4 rounded-full"
+                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
                             >End Proposals Registering</button>
-                        ) : (<></>)}
+                        }
 
-                        {WorkflowStatus[workflowStatus] === "ProposalsRegistrationEnded" ? (
+                        {WorkflowStatus[workflowStatus] === "ProposalsRegistrationEnded" &&
                             <button
                                 onClick={() => updateWorkflowStatus("startVotingSession")}
-                                className="content-center bg-indigo-950 hover:bg-indigo-900 text-white font-semibold py-2 px-4 rounded-full"
+                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
                             >Start voting session</button>
-                        ) : (<></>)}
+                        }
 
-                        {WorkflowStatus[workflowStatus] === "VotingSessionStarted" ? (
+                        {WorkflowStatus[workflowStatus] === "VotingSessionStarted" &&
                             <button
                                 onClick={() => updateWorkflowStatus("endVotingSession")}
-                                className="content-center bg-indigo-950 hover:bg-indigo-900 text-white font-semibold py-2 px-4 rounded-full"
+                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
                             >End voting session</button>
-                        ) : (<></>)}
+                        }
 
-                        {WorkflowStatus[workflowStatus] === "VotingSessionEnded" ? (
+                        {WorkflowStatus[workflowStatus] === "VotingSessionEnded" &&
                             <button
                                 onClick={() => updateWorkflowStatus("tallyVotes")}
-                                className="content-center bg-indigo-950 hover:bg-indigo-900 text-white font-semibold py-2 px-4 rounded-full"
+                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
                             >Tally votes</button>
-                        ) : (<></>)}
+                        }
                     </div>
 
                 </div>
-            ) : (<></>)}
-
-            {success && <div className="text-green-600 font-semibold p-4 text-center">Workflow status updated! Transaction: {success}</div>}
-            {error && <div className="text-red-800 font-semibold p-4 text-center">{error}</div>}
+            }
 
             <Event name='WorkflowStatusChange'></Event>
         </>